Move feature record functions into a separate module
