+++ ipfx/feature_record.py
@@ -1,0 +1,136 @@
+import pandas as pd
+import numpy as np
+
+
+def build_cell_feature_record(cell_features, sweep_features):
+    ephys_features = {}
+
+    # find hero and rheo sweeps in sweep table
+    rheo_sweep_num = cell_features["long_squares"]["rheobase_sweep"]["id"]
+    rheo_sweep_id = sweep_features.get(rheo_sweep_num, {}).get('id', None)
+
+    if rheo_sweep_id is None:
+        raise Exception("Could not find id of rheobase sweep number %d." % rheo_sweep_num)
+
+    hero_sweep = cell_features["long_squares"]["hero_sweep"]
+    if hero_sweep is None:
+        raise Exception("Could not find hero sweep")
+
+    hero_sweep_num = hero_sweep["id"]
+    hero_sweep_id = sweep_features.get(hero_sweep_num, {}).get('id', None)
+
+    if hero_sweep_id is None:
+        raise Exception("Could not find id of hero sweep number %d." % hero_sweep_num)
+
+    # create a table of values
+    # this is a dictionary of ephys_features
+    base = cell_features["long_squares"]
+    ephys_features["rheobase_sweep_num"] = rheo_sweep_num
+    ephys_features["thumbnail_sweep_num"] = hero_sweep_num
+    ephys_features["vrest"] = nan_get(base, "v_baseline")
+    ephys_features["ri"] = nan_get(base, "input_resistance")
+
+    # change the base to hero sweep
+    base = cell_features["long_squares"]["hero_sweep"]
+    ephys_features["adaptation"] = nan_get(base, "adapt")
+    ephys_features["latency"] = nan_get(base, "latency")
+
+    # convert to ms
+    mean_isi = nan_get(base, "mean_isi")
+    ephys_features["avg_isi"] = (mean_isi * 1e3) if mean_isi is not None else None
+
+    # now grab the rheo spike
+    base = cell_features["long_squares"]["rheobase_sweep"]["spikes"][0]
+    ephys_features["upstroke_downstroke_ratio_long_square"] = nan_get(base, "upstroke_downstroke_ratio")
+    ephys_features["peak_v_long_square"] = nan_get(base, "peak_v")
+    ephys_features["peak_t_long_square"] = nan_get(base, "peak_t")
+    ephys_features["trough_v_long_square"] = nan_get(base, "trough_v")
+    ephys_features["trough_t_long_square"] = nan_get(base, "trough_t")
+    ephys_features["fast_trough_v_long_square"] = nan_get(base, "fast_trough_v")
+    ephys_features["fast_trough_t_long_square"] = nan_get(base, "fast_trough_t")
+    ephys_features["slow_trough_v_long_square"] = nan_get(base, "slow_trough_v")
+    ephys_features["slow_trough_t_long_square"] = nan_get(base, "slow_trough_t")
+
+    ephys_features["threshold_v_long_square"] = nan_get(base, "threshold_v")
+    ephys_features["threshold_i_long_square"] = nan_get(base, "threshold_i")
+    ephys_features["threshold_t_long_square"] = nan_get(base, "threshold_t")
+    ephys_features["peak_v_long_square"] = nan_get(base, "peak_v")
+    ephys_features["peak_t_long_square"] = nan_get(base, "peak_t")
+
+    base = cell_features["long_squares"]
+
+    ephys_features["sag"] = nan_get(base, "sag")
+    # convert to ms
+    tau = nan_get(base, "tau")
+    ephys_features["tau"] = (tau * 1e3) if tau is not None else None
+    ephys_features["vm_for_sag"] = nan_get(base, "vm_for_sag")
+    ephys_features["f_i_curve_slope"] = nan_get(base, "fi_fit_slope")
+
+    # change the base to ramp
+    base = cell_features["ramps"]["mean_spike_0"] # mean feature of first spike for all of these
+    ephys_features["upstroke_downstroke_ratio_ramp"] = nan_get(base, "upstroke_downstroke_ratio")
+    ephys_features["peak_v_ramp"] = nan_get(base, "peak_v")
+    ephys_features["peak_t_ramp"] = nan_get(base, "peak_t")
+    ephys_features["trough_v_ramp"] = nan_get(base, "trough_v")
+    ephys_features["trough_t_ramp"] = nan_get(base, "trough_t")
+    ephys_features["fast_trough_v_ramp"] = nan_get(base, "fast_trough_v")
+    ephys_features["fast_trough_t_ramp"] = nan_get(base, "fast_trough_t")
+    ephys_features["slow_trough_v_ramp"] = nan_get(base, "slow_trough_v")
+    ephys_features["slow_trough_t_ramp"] = nan_get(base, "slow_trough_t")
+
+    ephys_features["threshold_v_ramp"] = nan_get(base, "threshold_v")
+    ephys_features["threshold_i_ramp"] = nan_get(base, "threshold_i")
+    ephys_features["threshold_t_ramp"] = nan_get(base, "threshold_t")
+
+    # change the base to short_square
+    base = cell_features["short_squares"]["mean_spike_0"] # mean feature of first spike for all of these
+    ephys_features["upstroke_downstroke_ratio_short_square"] = nan_get(base, "upstroke_downstroke_ratio")
+    ephys_features["peak_v_short_square"] = nan_get(base, "peak_v")
+    ephys_features["peak_t_short_square"] = nan_get(base, "peak_t")
+
+    ephys_features["trough_v_short_square"] = nan_get(base, "trough_v")
+    ephys_features["trough_t_short_square"] = nan_get(base, "trough_t")
+
+    ephys_features["fast_trough_v_short_square"] = nan_get(base, "fast_trough_v")
+    ephys_features["fast_trough_t_short_square"] = nan_get(base, "fast_trough_t")
+
+    ephys_features["slow_trough_v_short_square"] = nan_get(base, "slow_trough_v")
+    ephys_features["slow_trough_t_short_square"] = nan_get(base, "slow_trough_t")
+
+#    print "slow trough t:", ephys_features["slow_trough_t_short_square"]
+
+    ephys_features["threshold_v_short_square"] = nan_get(base, "threshold_v")
+    #ephys_features["threshold_i_short_square"] = nan_get(base, "threshold_i")
+    ephys_features["threshold_t_short_square"] = nan_get(base, "threshold_t")
+
+    ephys_features["threshold_i_short_square"] = nan_get(cell_features["short_squares"], "stimulus_amplitude")
+
+    return ephys_features
+
+
+def build_sweep_feature_records(sweep_table, sweep_features):
+
+    sweep_table = sweep_table.copy()
+
+    num_spikes = []
+    pds = []
+    for sn in sweep_table['sweep_number']:
+        sweep = sweep_features.get(sn, {})
+        pds.append(sweep.get('peak_deflect', [None])[0])
+        num_spikes.append(len(sweep.get('spikes',[])))
+
+    sweep_table['peak_deflection'] = pd.Series(pds)
+    sweep_table['num_spikes'] = pd.Series(num_spikes)
+
+    return sweep_table.to_dict(orient='records')
+
+
+
+def nan_get(obj, key):
+    """ Return a value from a dictionary.  If it does not exist, return None.  If it is NaN, return None """
+    v = obj.get(key, None)
+
+    if v is None:
+        return None
+    else:
+        return None if np.isnan(v) else v
